Extract chunk duration publisher
