+++ Sources/Player/Player.swift
@@ -458,17 +458,7 @@ extension Player {
     }
 
     private func configureChunkDurationPublisher() {
-        rawPlayer.publisher(for: \.currentItem)
-            .map { item -> AnyPublisher<CMTime, Never> in
-                guard let item else { return Just(.invalid).eraseToAnyPublisher() }
-                return item.asset.propertyPublisher(.minimumTimeOffsetFromLive)
-                    .map { CMTimeMultiplyByRatio($0, multiplier: 1, divisor: 3) }       // The minimum offset represents 3 chunks
-                    .replaceError(with: .invalid)
-                    .prepend(.invalid)
-                    .eraseToAnyPublisher()
-            }
-            .switchToLatest()
-            .removeDuplicates()
+        rawPlayer.chunkDurationPublisher()
             .receiveOnMainThread()
             .lane("player_chunk_duration")
             .assign(to: &$chunkDuration)
