+++ Tests/PlayerTests/ChunkDurationPublisherTests.swift
@@ -1,0 +1,52 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+@testable import Player
+
+import AVFoundation
+import Circumspect
+import XCTest
+
+final class ChunkDurationPublisherTests: XCTestCase {
+    func testChunkDuration() {
+        let item = AVPlayerItem(url: Stream.shortOnDemand.url)
+        let player = AVPlayer(playerItem: item)
+        expectEqualPublished(
+            values: [.invalid, CMTime(value: 1, timescale: 1)],
+            from: player.chunkDurationPublisher(),
+            during: 3
+        )
+    }
+
+    func testChunkDurationDuringEntirePlayback() {
+        let item = AVPlayerItem(url: Stream.shortOnDemand.url)
+        let player = AVPlayer(playerItem: item)
+        expectAtLeastEqualPublished(
+            values: [.invalid, CMTime(value: 1, timescale: 1)],
+            from: player.chunkDurationPublisher()
+        ) {
+            player.play()
+        }
+    }
+
+    func testCheckDurationsDuringItemChange() {
+        let item1 = AVPlayerItem(url: Stream.shortOnDemand.url)
+        let item2 = AVPlayerItem(url: Stream.onDemand.url)
+        let player = AVQueuePlayer(items: [item1, item2])
+        expectEqualPublished(
+            values: [
+                .invalid,
+                CMTime(value: 1, timescale: 1),
+                .invalid,
+                CMTime(value: 4, timescale: 1)
+            ],
+            from: player.chunkDurationPublisher(),
+            during: 3
+        ) {
+            player.play()
+        }
+    }
+}
